version 1.0 - Completed project.
Added in date comparison class for comparing number of days, months
and years between 2 dates provided by the user.

Updated the while loops to be only 1, removing 2nd while loop
making this more efficient.

Added in prints, showing days, months and years.

Added in main menu and used while loops to allow user to confirm
if they want to calculate the above between both dates and to
close off, as necessary. Double checked the error messages and
they look good.

Because of this I only needed to keep it to 1 file.